Added Edit Message Functionality
Refactored "remove" functionality to take list index, instead of
computing list index based on text matching. This allows less callback
binding and relying on the "this" keyword.

"Edit" functionality is essentially the same as the "remove"
functionality, but additionally assigning the given messages value to
tempTxt to be rendered in the MessageInput & DisplayText components.

Refactored directory structure to differentiate between "dumb"
components that are only concerned with rendering out data pumped into
them, and "smart" containers that evaluate/process data and state.

Changes to be committed:

-	modified:   app/javascript/bundles/App/actions/index.js
-	new file:   app/javascript/bundles/App/components/MessageList.js
-	new file:   app/javascript/bundles/App/components/TextDisplay.js
-	deleted:    app/javascript/bundles/App/containers/EditTextDisplay.js
-	new file:   app/javascript/bundles/App/containers/MessageInput.js
-	deleted:    app/javascript/bundles/App/containers/MessageList.js
-	new file:   app/javascript/bundles/App/containers/Messages.js
-	modified:   app/javascript/bundles/App/containers/SomethingSmall.js
-	deleted:    app/javascript/bundles/App/containers/TextInputControl.js
-	modified:   app/javascript/bundles/App/reducers/basic.js
-	modified:   app/javascript/bundles/App/startup/BasicFlow.js